requirements.txt
# Risk service dependencies
fastapi==0.104.1
pydantic==2.5.0
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
prometheus-client==0.19.0

# Reporting
WeasyPrint==60.1
ReportLab==4.0.7
plotly==5.18.0
kaleido==0.2.1

# Logging & monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Utils
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2
